+++ docs/src/pages/framed/DatePicker/DatePickerModal.svelte
@@ -2,7 +2,12 @@
   import { Modal, DatePicker, DatePickerInput } from "carbon-components-svelte";
 </script>
 
-<Modal open modalHeading="DatePicker">
+<Modal
+  open
+  modalHeading="Meeting date"
+  primaryButtonText="Confirm"
+  secondaryButtonText="Cancel"
+>
   <DatePicker datePickerType="single" flatpickrProps="{{ static: true }}">
     <DatePickerInput labelText="Meeting date" placeholder="mm/dd/yyyy" />
   </DatePicker>