+++ docs/src/pages/framed/Tabs/TabsReactive.svelte
@@ -8,11 +8,11 @@
   <Tab label="Tab label 1" />
   <Tab label="Tab label 2" />
   <Tab label="Tab label 3" />
-  <div slot="content">
+  <svelte:fragment slot="content">
     <TabContent>Content 1</TabContent>
     <TabContent>Content 2</TabContent>
     <TabContent>Content 3</TabContent>
-  </div>
+  </svelte:fragment>
 </Tabs>
 
 <div>