+++ CHANGELOG.md
@@ -7,6 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+## [0.51.0](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.51.0) - 2021-12-30
+
+**Features**
+
+- enhance `Checkbox` to support `bind:group`
+
+## [0.50.5](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.50.5) - 2021-12-30
+
+**Fixes**
+
+- filterable `MultiSelect` should blur when pressing "Tab"
+- filterable `MultiSelect` menu should close when pressing "Escape"
+- filterable `MultiSelect` menu should open when pressing "Space"
+- `DatePicker` input should not lose focus when pressing "Enter"
+- `ToolbarSearch` should expand if `value` is programmatically set
+
+## [0.50.4](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.50.4) - 2021-12-30
+
+**Fixes**
+
+- pressing "Escape" in a `DatePicker` should not propagate past the open calendar
+- prevent menu icon in filterable `MultiSelect` from triggering twice
+
+## [0.50.3](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.50.3) - 2021-12-28
+
+**Fixes**
+
+- do not dispatch "click", "click:row" `DataTable` events when selecting a checkbox or radio button
+- render "labelText" slot if `labelText` prop or `$$slots.labelText` is truthy
+- prevent default behavior in `Tabs` mobile variant when clicking a menu option
+- coalesce nullish values to `""` for `TextInput`, `PasswordInput`, `TextArea`
+
 ## [0.50.2](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.50.2) - 2021-11-27
 
 **Fixes**